[NF] generalization of oct3filter to nth oct
