+++ tests/nth_oct/tst_nth_oct.py
@@ -4,11 +4,6 @@ import numpy as np
 from mosqito.functions.shared.load import load
 from mosqito.functions.oct3filter.oct3spec import oct3spec
 from mosqito.functions.oct3filter.comp_noct_spectrum import comp_noct_spectrum
-from mosqito.functions.noctfilter.n_oct_filter import (
-    getFrequencies,
-    designFilters,
-    analyseData,
-)
 
 signal, fs = load(
     "./validations/loudness_zwicker/data/ISO_532-1/Test signal 5 (pinknoise 60 dB).wav",
@@ -17,13 +12,7 @@ signal, fs = load(
 spec_third, third_axis = oct3spec(signal, fs)
 plt.semilogx(third_axis, spec_third, label="oct3spec")
 
-f_dict = getFrequencies(24, 12600, 3)
-filters = designFilters(f_dict, fs, plot=False)
-rmsData = analyseData(filters, signal, f_dict, plot=False)
-spec_dB = 20 * np.log10((rmsData) / (2 * 10 ** -5))
-plt.semilogx(third_axis, spec_dB, label="noctfilter")
-
-spec_1, freq = comp_noct_spectrum(signal, fs, 24, 12600)
+spec_1, freq = comp_noct_spectrum(signal, fs, 24, 12600, n=3)
 spec_1_dB = 20 * np.log10((spec_1) / (2 * 10 ** -5))
 plt.semilogx(freq, spec_1_dB, label="comp_noct_spectrum")
